cros-ide: Fix VNC test flakiness
Install an error handler to sockets in the fake SSH server. Otherwise
socket errors (like EPIPE, which is quite usual) will make tests fail
for uncaught exceptions.

Also find an unused port in a valid manner so that the VNC test doesn't
fail when it tries to use a port already in use.

BUG=b:233965775
TEST=npm test  # with a local mod to run VNC test 100 times

Change-Id: Ie470201cb2621e33d06b9494733b674fcfca099d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3670638
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>