cros-ide: Only link to non-draft gerrit comments
Fixes an issue where draft comments were showing the 'link' icon in the
UI. When clicked, the user would be taken to a 404 page because the
comment was not yet linkable within Gerrit. To match Gerrit's
functionality, links to comments in the IDE now only appear for
published comments.

BUG=b:266773974
TEST=https://screenshot.googleplex.com/6sR5xiNqsQjR2NN
TEST=https://screenshot.googleplex.com/3Vs8VysXP2Q6HfT
TEST=https://screenshot.googleplex.com/B4YoCZRbqZCcYR7

Change-Id: I85a1188e5a3b7d117364d76fb7944d9e47bd7499
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4225384
Commit-Queue: David Welling <davidwelling@google.com>
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: David Welling <davidwelling@google.com>