cros-ide: Add a centralized point for all the chromiumos services
The newly added class manages the lifetime of services depending on
chromeos source code or chroot, and notifies its update to the
subscribers. The device management feature will use this class so that
it can dynamically enable/disable commands depending on the existence
of ChrootService.

BUG=b:265958085
TEST=npm t

Change-Id: I5f053cde50c92f219e4e539a864f8a3878868abd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4248861
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>