cros-ide: Fix compilation failure on hermes package
Before this change, the code assumed all the 9999 ebuild files for
platform2 packages were under chromiumos-overlay. It is not true
for the packages including hermes whose ebuild files are in the
internal repository, hence their compilation failure.

This change fixes the issue by including the directory that contains
the internal ebuild files to the search path on finding the ebuild file
to compile the target.

BUG=b:236793033
FIXED=b:236793033
TEST=npm t
TEST=open hermes/hermes_common.cc and confirm code completion and
navigation works

Change-Id: I29cb08e1041faf224961d486b03f0982fb654dba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4560891
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>