cros-ide: Use partner testing rsa if available
This change updates the device management feature to additionally use
the parter testing rsa when the user is opening a directory under
internal chromeos repository as a workspace folder.

BUG=b:260009987
TEST=npm t
TEST=Lease a device, click the "Connect to device (shell)" button, and
observe the command IDE runs. Confirm the correct path to the partner
testing rsa is used in the option.

Change-Id: I535653acf843ececcfece5b1b452c5a86907905e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4578943
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>