cros-ide: Generate gtest filter based on test parametrization
This makes the gtest filter generation more accurate, so that a filter
for "Foo" does not include "BarFoo". To do so, we extend the parser to
check whether a test is parametrized, and only then generate `*/Test/*`
and `Test/*` filters. Otherwise we just use `Test` as the filter.

BUG=b:274425767
TEST=Ran and adjusted unit tests, and checked the filter generation in a
TEST=sample platform2 test.

Change-Id: I5768ccc68b635d618310131e9e216b54771e9673
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4629289
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Sonja Laurila <laurila@google.com>
Tested-by: Christian Flach <cmfcmf@chromium.org>
Commit-Queue: Christian Flach <cmfcmf@chromium.org>