ide: Introduce separate test items for test suites
We introduce a new `GtestSuite`, which sits between `GtestFile`s and
`GtestCase`s. As part of this, we also extract some common logic in an
abstract base class for all types of test item wrappers, which we call
`GtestRunnable`.

This also makes the test item ids shorter, because they only need to be
unique in their parent collection.

BUG=b:274425767
TEST=Run several combinations of test files, test suites, and test cases
TEST=in both Chromium and ChromiumOS.

Change-Id: Ib37c616c80fea59219c955f88d328f75bbc03c06
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4711591
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Sonja Laurila <laurila@google.com>
Tested-by: Christian Flach <cmfcmf@chromium.org>
Commit-Queue: Christian Flach <cmfcmf@chromium.org>