ide: Add a context menu to run regenerate compilation database
Usually compilation database is not regenerated when the last attempt
has failed. This change adds a context menu to the menu item notifying
the failure of compilation database generation, from which the user can
trigger recompilation of the active file. When it is clicked, we
attempt to trigger the regeneration even if it has failed before.

Before this change, the interface method shouldGenerate has been
returning boolean, but for this change, we have to distinguish at least
three states: 1. Should generate. 2. Unsupported, 3. Previous attempt
failed. For the case 3, we should regenerate the compdb only when the
user explicitly requests it. That's why we updated the method to return
an enum and accordingly classes implementing the interface and tests are
updated.

BUG=b:289851648
TEST=npm t
TEST=Open codelab.cc, save it, confirm the compdb has been generated,
click "ChromiumIDE", right click "C++ xrefs generation", click the menu
, and confirm the compdb is regenerated.

Change-Id: Id89c757de14731414bdf5629565c7a332c26db28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4665112
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>