ide: Add support for rotaeted VNC session
This patch adds "Connect to Device (VNC) with rotation"
as a new item in the context menu in "My Devices" section.

Context:
Frame buffers of devices like kukui are in portrait layout.
When those devices are used in landscape mode, VNC will
display the screen conetnts rotated 90 degrees, which is
quite confusing and difficult to use.

BUG=b:288590713
TEST=manually tested the context menu and it worked; npm test

Change-Id: If5d47a3feb580a3475d6a87f9cf821682e75ca04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4669632
Commit-Queue: Satoru Takabayashi <satorux@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: Satoru Takabayashi <satorux@chromium.org>