ide: Add command to run type check and run it on preupload command
Now that the commands to build and test the extension don't run type
check, this change adds a separate command `npm run check` to run type
check and updates `npm run preupload` to run this command. The preupload
command is not automatically run for now, but we will run it on CQ
eventually.

This change also updates tools/tsconfig.json to satisfy the type
checker.

BUG=None
TEST=npm run preupload

Change-Id: Ie6f1804b01af39f650d7afbad8fd7c8d44140995
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4737192
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>