ide: Allow tests to access fakes via testing.fakes.*.
This change exports the fakes module under testing/ in testing/index.ts.
This allows the test code to access utility functions in the fakes
module in the form of testing.fakes.*. This is useful because tests
typically import testing, and arguably this change will make the fakes
module more discoverable for test authors, making the module appears on
IDE suggestion when the author types `testing.`.

Changes outside index.ts are needed to break cyclic dependencies that
are disallowed by our linter for correctness of build.

BUG=None
TEST=npm run preupload

Change-Id: I6e11c1d9e7641eca04f29b1994ad7858d3465bb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4828192
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>