ide: gerrit: continue on failing to fetch user's account info
This change allows the user to see public Gerrit comments even if they
don't have credential to upload a Gerrit patch. Given the change allows
the user to see the UI to add a reply comment without credentials, this
change also makes ChromiumIDE show an error message when the user tries
to add a reply and it fails.

BUG=b:296808152
TEST=comment out the line starts with chromium-review.googlesource.com
in ~/.gitcookies and confirm a public comment is shown.
TEST=Then try to add a draft reply to the comment and confirm an error
message is shown.
TEST=Then click the error message and confirm the website for Gerrit
credential setup is opened.

Change-Id: I144cca332e8ec5ffaa9f143089bc0da692cb4d0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4841476
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>