ide: Add option to tree-kill processes and enable it for autoninja
This is a new option for commonUtil.exec to tree-kill a process and all
of its subprocesses instead of just the process itself. I noticed that
autoninja was not properly killed when cancelling a test run, because
subprocesses are not killed by default when calling `command.kill()`.

BUG=b:301574822
TEST=Wrote a unit test and verified that autoninja is killed correctly.

Change-Id: I30db70abab43cc8b18aaf844623410e2b79fef94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4887451
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Christian Flach <cmfcmf@chromium.org>
Tested-by: Christian Flach <cmfcmf@chromium.org>