ide: Extract the function to ensure SSH session
This change extracts out the function to ensure SSH session as a common
utility and makes it report failure. For the function not to fail in
unit tests, we expand our fake SSH server implementation to be able to
handle the requests the function makes with the real SSH command.

The extracted function will be used by a command to deploy a package to
a DUT that will be added.

BUG=b:214336231
BUG=b:301891679
TEST=npm run preupload

Change-Id: I63ee62afaaabe1bb1f41887f8d78c0147e014301
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4891414
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>