ide: boards and packages: Add build command
This change adds a context menu to the boards and packages view to build
the selected package for the selected board.

This is a part of the changes to overhaul boards and packages view.
See the commit message of https://crrev.com/c/4791430 for all the TODOs
we will fix through these changes.

Design doc: http://go/chromiumide-dd-build-and-deploy

BUG=b:214336231
TEST=npm run preupload

Change-Id: Ie5b48c79d48204e073e9044f266f9c397561435d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4900221
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>