ide: Don't reveal packages if tree view is not visible
Before this change, when a file belonging to a chromeos package was
revealed chromiumIDE would always reveal the package in the boards and
packages view. After this change, we do this only when the boards and
packages tree view is already visible, so that the view that is opened
in the view container (e.g. file explorer) doesn't change.

BUG=b:310765143
TEST=manual
TEST=npm t

Change-Id: I244120c150f6eb2bc1fa467344b3e7201c0b1bd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/5028985
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>