ide: add command to deploy package to device
User could click on package icon on device management panel (next to the
flash image icon) and choose package to deploy from dropdown. Package
that has been `cros-workon start` are on top of list for ease.
This is a simple wrapper of the `cros deploy` command.

BUG=b:308059094
TEST=npm t
TEST=(icon) https://screenshot.googleplex.com/BGJYdFmTryTivKE
TEST=(choose package) https://screenshot.googleplex.com/3vjY7mVWZcqXsji
TEST=(succeeded) https://screenshot.googleplex.com/59QuMGhVouqjQCN

Change-Id: Ibaddc5da52033ed40a87b5e8e7e0e5a3de730b93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5076036
Tested-by: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>