ide: Remove dead code around crosfleet lease
The return value of the requestLease method was added to implement
device manager V2 before, but the contributor is no longer working on
this project and we don't have plan to rework on it for the time being.
This change removes the dead code around the logic. It helps us
understand what aspects of crosfleet CLI are actually used from
ChromiumIDE.

As additional clean up while we are here, the method is renamed to
indicate that it throws.

BUG=b:311558537
TEST=npm t

Change-Id: I376968a713f6aa0582243aa7a5003dc6971841a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5086249
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>