ide: ensure flushMicrotasksUntil to clean up resources
Before this change, the test utility flushMicrotasksUntil would keep
polling if it finishes due to timeout, or keep setTimeout job in the
event queue if it finishes due to the condition being satisfied. This
change fixes those inefficiency issues.

This is a clean up change which have no visible impact, given the number
of usages of this function is small.

BUG=None
TEST=npm t

Change-Id: Ia55d5e12c0b341fb02d4ca2118364d3c1d3f3b50
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5118900
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>