ide: device manager panel shows device info (board, model, builder path)
Make device details (builder path, and board for owned devices) more
visible so that user can see get a glance the state a device is in and
determine if they want to flash it with a new image.
Since the 'description' field of device item is already too long to be
displayed fully in the side panel, make each property a child items of
the device item, and provide a copy command on right click.

Note that model is only available for leased devices (not provided by
/etc/lsb-release).

BUG=b:308059094, b:241846471
TEST=https://screenshot.googleplex.com/4fdGsHjHTy5QmAH

Change-Id: I99485f14b64fcd0e689b7f319dadb14e27b7b820
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5098787
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>