ide: device management: use pre-fetched device board information
Pass deviceClient instance to CommandContext for the flashing image to
device and deploying package commands, so that user do not have to wait
for IDE to fetch device metadata (board) on-the-fly to proceed to image
and package selection.

There is now zero delay from clicking "flash image" icon to image
type selection.
Similarly, it is almost instantaneous for "deploy package" to show the
list of packages available together with image caching.

BUG=b:308059094
TEST=npm t
TEST=flash image: https://screenshot.googleplex.com/MDErFeyU7JLt4fR
TEST=deploy package: https://screenshot.googleplex.com/7pbaGnwfU6ttZV4

Change-Id: Ib60a3f8013f2c91ac124df14ab346251078e440a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5101752
Tested-by: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>