device: add function to suggest image based on check config/input
Similar to the general flash image menu, user will first be asked to
choose from a list of image types (filtered to those compatible with the
cros-debug flag status), then the image from a list of them with
different versions.

Since the `gsutil ls` takes considerable time to run and does not
support regex (to search for images with different CrOS major version in
one command), by default show only those with the exact same one as the
local version, and expand the list to all within the range (+/- config
threshold) if user chooses to.
If no image matches the criteria, provide option to choose another image
type. This is quite likely if user has an older repository and chooses
postsubmit images which are only retained for 5 days.

Move utility functions for constructing Checker* to
test/unit/.../check_image/common.ts so it could be used for unit tests
of both compatibility checker and suggest image function.

This is not used anywhere yet (to be added to image check in next CL).

BUG=b:308059094
TEST=npm t
TEST=choose image type: \
     https://screenshot.googleplex.com/BjoKV4YzFJaGxSw
TEST=default image same as local version: \
     https://screenshot.googleplex.com/3MzwucGvaMXcfnr
TEST=loading more images: \
     https://screenshot.googleplex.com/4Xd8DY2Q4kN2vwQ \
     https://screenshot.googleplex.com/5yzcMc9sS6AaCri

Change-Id: Ibea49feeaf53209fa81538997b9857a9d1b01ca4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5294359
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>