image check: add display result in non-modal message or quickpick option
The image check function would typically be run as a background step or
a part of another device feature (deploy), and modal message is not
suitable for such purposes.

Add the two new modes which only report failed checks, and in a less
intrusive way (non-modal message could be ignored) or integrated with
the rest of the command (deploy uses quickpick for device and
package selection).

BUG=b:308059094
TEST=None

Change-Id: I0372fa3ec4835930a25a1a5dbff746631f7ddfcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5320859
Tested-by: Grace Cham <hscham@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>