Move cros format to shared/app/
Previously, the feature is activated only when inside a CrOS repository.
This CL changed it to always be triggered (if enabled in config), and do
nothing if `cros` is not found (i.e. not in a CrOS cehckout).

Note this aligns with cros lint's behavior.

BUG=b:319548749
TEST=npm t
TEST=manual on markdown file (trailing whitespace)

Change-Id: I43bf1438ef5b270409537df0c2b2bb08abc60b95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5427375
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>