metrics: wrap in driver layer and move to shared/
Expose only the activate (called when shared features are activated) and
send functions which are the only ones used outside of the module.
We need a different activate function on the two platforms because of
the different privacy requirements.

The change is large but many are automatic/mechanically renaming
`driver.sendMetrics` to `driver.metrics.send` and similarly for the spy
in tests.

BUG=b:322256481
TEST=metrics events showing up on testing GA realtime dashboard
TEST=resetting metrics collection consent form works
TEST=npm t

Change-Id: Ib8da224aca2e1733333d91f4688eddd7ad7c3824
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5504889
Commit-Queue: Grace Cham <hscham@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>