metrics: Detect gMac to send metrics
Enable metrics on gMac as well, so that we can send some metrics in case
the user is somehow installing chromiumide on Mac.

BUG=b:333818718
TEST=npm t
TEST=Build .vsix file on linux (npx vsce package -o /tmp/foo.vsix), copy
the file to Mac, install the .vsix file on Mac, and confirm the warning
is shown. Then I opened GA to see realtime view and confirmed an event
for activation has os = Darwin attribute. (tested on https://crrev.com/c/5504895)

Change-Id: Iadb1710435503903864a90f912f008e6e4abaff1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5504894
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>