Refactor WrapFs for simpler migration to shared/
To prepare for migrating the code to shared/, the class will switch to
use driver.fs instead of nodejs for compatibility with cider.
This change simplifies that by removing unused functions in WrapFs.

Due to limitations of cider's fs APIs, where fs.Stats in not an
available type, and the closest type from vscode does not have aTime,
split the `WrapFs.stat` function to getting aTime and mTime which are
its only usages.
Although we will not be able to implement the `WrapFs.aTime` function on
cider side, it is OK because it will not be used there and can just
throw a 'not implemented' error.

BUG=None
TEST=npm t

Change-Id: Ia1c9cd86070610ccd07f9f7d5c34373e039fb2ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5524484
Auto-Submit: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>