ebuild: Implement document link provider in LSP
Tests are redirected to test the LSP implementation to ensure that is
feature parity with the existing implementation.

With this CL, the features LSP provide are on per with the existing
features and we are ready to flip the flag to make LSP the default.

The logic in the document_links.ts file is copied from
src/features/chromiumos/ebuild/link_provider.ts. The duplication will be
removed soon after the migration completes.

BUG=b:343112937
TEST=npm t

Change-Id: I620f5fa9bd0b9da6a2aee851d109bde1e45a1cd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5587780
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>