ui: Move showInputBoxWithSuggestions under ui/
This is a pure refactoring. A parameter is renamed and got its type
updated allowing the caller to specify descriptions, etc. not only the
labels.

We will improve the component and use it from a boards and packages command.

BUG=b:344786569
TEST=npm t

Change-Id: Idab57fe5461d51779165a32565f500d57e928f48
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5599859
Commit-Queue: Grace Cham <hscham@chromium.org>
Auto-Submit: Keigo Oka <oka@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>