ui: add new vscode.QuickPickItem types for item with (prefill) buttons
For some usages, the preset items might serve as a template that users
would edit and use the modified input as final accepted value.
One example is adding owned device - the list fetched from ssh config
contains the hostname but user might want to add a specific port.

If items are of the new `QuickPickItemWithPrefillButton` type
(initialized exactly like the existing `SimplePickItem`), the button
will be added automatically.
`showInputBoxWithSuggestions` is updated to handle the event that a
prefill button is clicked. User triggers it by either by a mouse click,
or tab twice and hit enter while the item is active.

See next change that converts the callsite for adding owned device to
use this new type.

BUG=b:344786569
TEST=no change to add owned device command input box

Change-Id: Ic75b16cdfd77e4a492a271f8dfb0ef721900237a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5599862
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>