format: Honor PRESUBMIT.cfg
Use cros format command that is defined in PRESUBMIT.cfg if any as a
base on constructing the format command to run.

OptionsParser is updated to handle long options and positional args, and
moved to shared/ so it can be used from formatter.

BUG=b:337139742
FIXED=b:337139742
TEST=npm t

Change-Id: I29b2cdf2dc2a487f4f90dd93828c396a418bcfcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5612792
Reviewed-by: Grace Cham <hscham@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>