lint: Refactor the file structure
This is a pure refactoring, as a preparation for a change to honor
PRESUBMIT.cfg on running cros lint.

Rename cros_lint -> lint as the feature also can run tast lint and
libchrome lint.

Split files per functionalities. Extract LintConfig interface and let
its implementations have a concrete class name.

GoLintConfig inherits the confusion the original code had that it
handles go lint and tast lint in a single function. We could refactor
the logic to make more sense, but that would make this change not a pure
refactoring.

BUG=b:337138396
TEST=npm t

Change-Id: If7e95b15c98461830a7ba88cae711cfd4e3b2ed3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5619495
Reviewed-by: Grace Cham <hscham@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>