Specify --cache on linting code with eslint
The gts command is a simple wrapper of the eslint command [1], but it
doesn't expose eslint's --cache flag to its CLI [2], which lets the tool
check only the changed files, a desired option to speed up our
iteration.

Before this change `npm run lint-code` took 44.2s. Now it's 3s if the
command is run after only 1 file is changed, if the command is run
before the change.

[1] https://github.com/google/gts/blob/main/src/cli.ts#L131C16-L136C40
[2] https://github.com/google/gts/issues/659

BUG=b:346849057
TEST=Run npm run lint-code, change a TS file to have an issue, run npm
run lint-code again and confirm the issue is detected.

Change-Id: I3da1865f113dc4c25b3668c78f280e3151afa3c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5632515
Reviewed-by: Grace Cham <hscham@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>