findGitDir: take optional arg root dir where search should terminate
The arg will be used in ChromiumOS-specific callsites where the search
should stop at the ChromiumOS root (and return undefined).

This is a refactoring with a small behavior change with the default
function when root is not specified:
* before: /.git will be checked and returned if exist
* now: a directory in / will be the last directory checked
This is considered a bug fix since root directory cannot be a git repo
directory (and typical usage is to be given the ChromiumOS root path
which is also not a git repo itself).

BUG=b:346900341
TEST=npm t

Change-Id: I4f4cc6c2d0173c18437befa33fbbb00852cb17c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5632521
Commit-Queue: Grace Cham <hscham@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>