cros lint: exec command with default cwd and resolved {ex,in}cludes
On cider commands running on a different cwd than the one allowed in the
security filter will be forbidden immediately.
It would be impossible to list out (and add the command entry) for all
CrOS repos, so instead we make the cros lint command ran with absolute
paths.

The test is taken from the original commit crrev.com/c/5623874.

BUG=b:322256659
TEST=Open infra/recipes as a workspace, open test_plan_filtering.py,
edit a `"""` to `"`, and confirm cros lint warning is shown. Then open
recipes.py, and confirm warnings from cros lint are not shown.
TEST=npm t

Change-Id: Idd2923cdd1f5ec60d88addd1b79474de66f42428
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5712169
Commit-Queue: Grace Cham <hscham@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>