ide: Allow multiple candidates for extension suggestions
For each language type, we can now specify multiple candidates of
extensions. A popup is shown to suggest the first candidate when none of
candidates are installed on the system.

Also adds "google.gn" to the extension candidates for GN. It's not
published yet, but coming soon.

BUG=None
TEST=npm t

Change-Id: I23eeadeaf4a93db82001198c40313c200b9da2e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/6178993
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>