ide: java/ds: Avoid scanning all classes on attach
VSCode tries to enable all breakpoints on attach, but the current logic
for it in the debug server is very slow because it needs to query all
classes loaded in the VM (it takes ~5ms per class on my environment).

This patch workarounds the problem by skipping classes that are clearly
irrelevant.

BUG=b:385660974
TEST=attach to an Android process

Change-Id: I9d4ed9988f02d4e87d6694d6b167553a8de13cfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/6193070
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>