ide: java/ds: Fix breakpoints in classes with inner/anonymous classes
Existing breakpoint installation logic assumes that there is one-to-one
correspondence between a class and a Java source file. But it does not
hold for classes with inner classes and/or anonymous classes.

This patch fixes the logic to handle the case. Also adds a regression
test.

BUG=b:394008365
TEST=mvn test -Dtest=JavaDebugServerTest

Change-Id: Ie52e50715beeee53b17461f8a76e4e47aa5744b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/6228495
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>